+++ services/table-geocoder/lib/table_geocoder.rb
@@ -169,7 +169,7 @@ module CartoDB
     end
 
     def import_results_to_temp_table
-      connection.copy_into(temp_table_name.lit, data: File.read(deflated_results_path), format: :csv)
+      connection.copy_into(Sequel.lit(temp_table_name), data: File.read(deflated_results_path), format: :csv)
     end
 
     def load_results_into_original_table